LayoutTests/editing/deleting/delete-list-items-in-table-cell-1-expected.txt

When deleting the second ordered list items in a table cell its content is merged with the first list item:

Before:
| <table>
|   <tbody>
|     <tr>
|       <td>
|         <ol>
|           <li>
|             "a"
|           <li>
|             id="li"
|             "<#selection-caret>|b"

After:
| <table>
|   <tbody>
|     <tr>
|       <td>
|         <ol>
|           <li>
|             "a<#selection-caret>|b"
